tests: the MISRA mutation gate runs against this fork's tree
Two environment gaps and one finding, all three reachable from the documented standalone
invocation (`PYTHONPATH=../opendbc_repo .venv/bin/python -m unittest discover -s tests`).

`test_misra.sh` calls bare `python3` for opendbc's include path and for comma's cppcheck, and
the copy the harness runs in excludes `.venv` — so under an activated venv that name is the
right interpreter, and under ours it is the system one, which has neither module. The harness
now puts its own interpreter's directory first. The same run needs the *fork's* opendbc:
`SConscript` reads its include path from an `import opendbc` of its own, and the child changes
directory, so the relative `PYTHONPATH` that works for the parent stops resolving and the build
compiles against the venv's installed upstream copy — where `controls_allowed_lateral` does not
exist. The harness now passes this run's `opendbc.INCLUDE_PATH` down, absolutely.

With the check finally looking at the right tree it reports a real one: the alternate-function
mask in `gpio.h`, `~(0xFU << ((pin & 7U) * 4U))`. MISRA 12.2 reads `0xFU`'s essential type as 8
bits wide while the shift reaches 28, and whether cppcheck reports it at all depends on what
else is in the translation unit — the new safety layer is what surfaced it here, and nothing
about the line is specific to this fork. Casting the literal to `uint32_t` gives the operand the
width the shift actually uses: same value, same result, and a check that is deterministic
instead of contingent. This tree's other gates and the health flag are untouched.

Verified: `python -m unittest discover -s tests` is 38 tests, OK, where before it was 37 with
one failure; the mutation cases that must be *caught* still are, so the gate kept its teeth, and
`scons` still builds both firmware targets.

+++ board/drivers/gpio.h
@@ -49,7 +49,11 @@ void set_gpio_output_type(GPIO_TypeDef *GPIO, unsigned int pin, unsigned int out
 void set_gpio_alternate(GPIO_TypeDef *GPIO, unsigned int pin, unsigned int mode) {
   ENTER_CRITICAL();
   uint32_t tmp = GPIO->AFR[pin >> 3U];
-  tmp &= ~(0xFU << ((pin & 7U) * 4U));
+  // moonpilot seam, see AGENTS.md: the mask literal cast to the width it is shifted within. MISRA
+  // 12.2 reads `0xFU`'s essential type as 8 bits and the shift can reach 28, which cppcheck reports
+  // or not depending on what else is in the translation unit; the cast makes the operand's width
+  // the one the shift actually uses. Same value, same result.
+  tmp &= ~((uint32_t)0xFU << ((pin & 7U) * 4U));
   tmp |= mode << ((pin & 7U) * 4U);
   register_set(&(GPIO->AFR[pin >> 3]), tmp, 0xFFFFFFFFU);
   set_gpio_mode(GPIO, pin, MODE_ALTERNATE);
